re-organize simulator and add edge simulation pipeline to readme
