+++ examples/advanced/edge/setup_system.md
@@ -1,6 +1,6 @@
 # Running Edge Example with Hierarchical Clients
 
-Please follow these steps to run the Edge device emulator,
+Please follow these steps to run the DeviceSimulator,
 
 ## Provision
 
@@ -8,7 +8,7 @@ Use tree_prov.py to generate a hierarchical NVFlare system with 2 levels and 2 c
 
      python nvflare/edge/tree_prov.py -r /tmp -p edge_example -d 1 -w 2
 
-This will create a deployment with 2 clients, 4 leaf-clients, 2 relays, 1 server.
+This will create a deployment with 4 leaf nodes, 2 aggregators, 2 relays, and 1 server.
 
 This file needs to be copied to the `local` folder of each leaf clients, C11, C12, C21 and C22.
 
@@ -55,16 +55,16 @@ print the aggregated results.
 
 The job can be started as usual in NVFlare admin console.
 
-## Run Edge Emulator
+## Run DeviceSimulator
 
-The emulator can be used to test all the features of the edge system. It handles 'train' task by simply doubling every values 
+The DeviceSimulator can be used to test all the features of the edge system. It handles 'train' task by simply doubling every values 
 in the weights.
 
-To start the emulator, give it an endpoint URL and number of devices like this,
+To start the DeviceSimulator, give it an endpoint URL and number of devices like this,
 
-    python /nvflare/edge/emulator/run_emulator.py http://localhost:8000 16
+    python /nvflare/edge/device_simulator/run_device_simulator.py http://localhost:8000 16
    
-The emulator keeps polling the LCP for job assignment. It only runs one job then quits.
+The DeviceSimulator keeps polling the LCP for job assignment. It only runs one job then quits.
