+++ examples/advanced/edge/simulators/binaries/train_cifar10
