+++ examples/advanced/edge/simulators/binaries/train_xor
