+++ examples/advanced/edge/simulators/configs/cifar10_et_cpp_config.json
@@ -1,7 +1,7 @@
 {
     "endpoint": "http://localhost:9003",
     "num_devices": 5,
-    "device_id_prefix": "emulator-",
+    "device_id_prefix": "sim-device-",
     "processor": {
       "path": "processors.et_cpp_task_processor.ETCppTaskProcessor",
       "args": {
