+++ examples/advanced/edge/simulators/configs/hello_config.json
@@ -1,7 +1,7 @@
 {
   "endpoint": "http://localhost:4321",
   "num_devices": 16,
-  "device_id_prefix": "emulator-",
+  "device_id_prefix": "sim-device-",
   "processor": {
     "path": "processors.hello_task_processor.HelloTaskProcessor",
     "args": {
