+++ examples/advanced/edge/simulators/configs/xor_pt_config.json
@@ -1,7 +1,7 @@
 {
   "endpoint": "http://localhost:8000",
   "num_devices": 16,
-  "device_id_prefix": "emulator-",
+  "device_id_prefix": "sim-device-",
   "processor": {
     "path": "processors.pt_xor_task_processor.PTXorTaskProcessor",
     "args": {
