+++ examples/advanced/edge/simulators/processors/et_cifar10_task_processor.py
