+++ examples/advanced/edge/simulators/processors/et_cpp_task_processor.py
@@ -20,7 +20,7 @@ import shutil
 import subprocess
 
 from nvflare.edge.constants import MsgKey
-from nvflare.edge.emulator.device_task_processor import DeviceTaskProcessor
+from nvflare.edge.device_simulator.device_task_processor import DeviceTaskProcessor
 from nvflare.edge.model_protocol import (
     ModelBufferType,
     ModelEncoding,
