+++ examples/advanced/edge/simulators/processors/et_xor_task_processor.py
